+++ drm/AampDRMSessionManager.h
@@ -312,7 +312,7 @@ public:
 	 *  @note		AccessToken memory is dynamically allocated, deallocation
 	 *				should be handled at the caller side.
 	 */
-	const char* getAccessToken(int &tokenLength, long &error_code ,bool bSslPeerVerify);
+	const char* getAccessToken(int &tokenLength, int &error_code ,bool bSslPeerVerify);
 	/**
 	 * @fn getDrmSession
 	 * @return index to the selected drmSessionContext which has been selected
